backend/requirements.txt
Django==5.2.13
djangorestframework==3.17.1
django-cors-headers==4.9.0
numpy==2.4.4
opencv-python-headless==4.13.0.92
